# Proyecto 1/Backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: SmartEgg API
  version: "2.0.0"
  description: |
    API REST de la plataforma IoT SmartEgg (Proyecto 1).

    Este documento define la interfaz que consume el frontend. Es la referencia
    unica de rutas, cuerpos, respuestas y codigos de estado.

    ## Separacion REST / MQTT

    | Va por REST (este documento) | Va por MQTT (broker) |
    |---|---|
    | CRUD de lotes, operadores y usuarios | Telemetria continua del nodo |
    | Consultas historicas y estadisticas | Alarmas criticas en tiempo real |
    | Login, roles y perfiles | Estado de actuadores |
    | Generacion de reportes PDF | Disponibilidad del nodo (LWT) |

    Las alarmas **no** se consultan por polling HTTP: el frontend se suscribe a
    `smartegg/{node_id}/alarmas/critica` por WebSocket (puerto 9001).
    El catalogo de topics vive en `GET /api/diagnostico/mqtt/topics`.

    ## Autenticacion

    1. `POST /api/auth/login` devuelve un JWT.
    2. Enviarlo en cada peticion protegida: `Authorization: Bearer <token>`.

    Roles: `administrador`, `operador_1`, `operador_2`.
    Un operador solo puede consultar **sus propios** diagnosticos e informes;
    el administrador puede consultar los de cualquiera.

    Los endpoints de telemetria de solo lectura son publicos a proposito, para
    que Grafana y las visualizaciones (p5.js / Chart.js) consuman datos reales.
servers:
  - url: http://localhost:5000
...
  - url: http://backend:5000
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Autenticacion]
      summary: Iniciar sesion y obtener el token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, format: password, example: admin123 }
      responses:
        "200":
          description: Credenciales correctas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RespuestaLogin" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401":
          description: Usuario o contrasena incorrectos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "403":
          description: La cuenta esta desactivada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
  /api/auth/yo:
    get:
      tags: [Autenticacion]
      summary: Datos del usuario dueno del token
      responses:
        "200":
          description: Usuario autenticado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
  /api/auth/cambiar-password:
    post:
      tags: [Autenticacion]
      summary: Cambiar la propia contrasena
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password_actual, password_nueva]
              properties:
                password_actual: { type: string, format: password }
                password_nueva:
                  type: string
                  format: password
                  minLength: 6
      responses:
        "200":
          description: Contrasena actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401":
          description: Sin token, o la contrasena actual no coincide
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
  /api/auth/usuarios:
    get:
      tags: [Autenticacion]
      summary: Listar usuarios
      description: Solo administrador. Nunca devuelve el hash de la contrasena.
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Usuario" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
    post:
      tags: [Autenticacion]
      summary: Crear un usuario
      description: Solo administrador.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, rol]
              properties:
                username: { type: string, example: operador3 }
                password: { type: string, format: password, minLength: 6 }
                rol: { $ref: "#/components/schemas/Rol" }
                nombre: { type: string, example: Maria Lopez }
                id_operador:
                  type: string
                  description: Obligatorio para los roles de operador.
                  example: OP3
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usuario" }
        "400": { $ref: "#/components/responses/ErrorValidacion" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
        "409": { $ref: "#/components/responses/Conflicto" }
        "503": { $ref: "#/components/responses/BaseNoDisponible" }
  /api/auth/usuarios/{username}:
    delete:
      tags: [Autenticacion]
      summary: Eliminar un usuario
      description: |
        Solo administrador. No se puede eliminar la propia cuenta ni la del
        administrador base definido en `ADMIN_USERNAME`.
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Usuario eliminado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mensaje" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "403": { $ref: "#/components/responses/SinPermisos" }
        "404": { $ref: "#/components/responses/NoEncontrado" }
        "409": { $ref: "#/components/responses/Conflicto" }

  # --------------------------------------------------------------- Perfil ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token de POST /api/auth/login. Cabecera: Authorization: Bearer <token>"
...
  responses:
    ErrorValidacion:
      description: El cuerpo de la peticion no paso las validaciones
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErroresValidacion" }
    NoAutenticado:
      description: Falta el token o no es valido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    SinPermisos:
      description: El rol del token no alcanza para esta operacion
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    NoEncontrado:
      description: El recurso no existe
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    Conflicto:
      description: El recurso ya existe o la operacion no esta permitida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
    BaseNoDisponible:
      description: |
        La base de datos no responde. El backend sigue en linea en modo
        degradado, pero no puede persistir.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Mensaje" }
  schemas:
    Mensaje:
      type: object
      properties:
        mensaje: { type: string }
    ErroresValidacion:
      type: object
      properties:
        errores:
          type: array
          items: { type: string }
          example: ["El campo 'codigo' es obligatorio."]
    Rol:
      type: string
      enum: [administrador, operador_1, operador_2]
    Usuario:
      type: object
      properties:
        username: { type: string, example: operador1 }
        nombre: { type: string, example: Operador 1 }
        rol: { $ref: "#/components/schemas/Rol" }
        id_operador:
          type: string
          nullable: true
          description: Enlace con el operador del nodo fisico. Null para el administrador.
          example: OP1
        activo: { type: boolean }
    RespuestaLogin:
      type: object
      properties:
        token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9... }
        token_tipo: { type: string, example: Bearer }
        expira_en: { type: string, format: date-time }
        usuario: { $ref: "#/components/schemas/Usuario" }
...